FFFR: Store primary contexts for multiple GPUs.
Instead of checking for a current context that may not match the requested gpu index instead always default to the primary context used by the runtime api.

This can also prevent an error in the case the current context actually is the primary context. We would then need to retain that context as otherwise it could be destroyed externally without fffr reatining a handle to it.